requirements - Copy.txt
# Core OCR Libraries
easyocr>=1.7.0
pytesseract>=0.3.10

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.26.0
scikit-image>=0.22.0

# Streamlit Web Interface
streamlit>=1.28.0

# Utilities
python-dateutil>=2.8.0

# Logging and Configuration
pyyaml>=6.0

# Optional: Advanced preprocessing
imutils>=0.5.4
